requirements.txt
# Core Flask dependencies (cross-platform)
flask==3.0.0
Flask-WTF==1.2.1
flask-talisman==1.1.0
flask-limiter==3.8.0

# AI/ML dependencies (cross-platform)
openai==1.12.0
httpx==0.27.0

# Document processing (cross-platform)
pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0

# File type detection (platform-specific)
python-magic==0.4.27; sys_platform != 'win32'
python-magic-bin==0.4.14; sys_platform == 'win32'

# HTTP client (cross-platform)
requests==2.32.3

# Configuration (cross-platform)
python-dotenv==1.0.1

# Security (cross-platform, pre-built wheels)
cryptography==44.0.0

# WSGI servers (platform-specific)
gunicorn==23.0.0; sys_platform != 'win32'
waitress==3.0.0; sys_platform == 'win32'
